Extract TestLog and LogTestCase into a header
Allow reusing them in other tests.

No changes yet, this is a pure refactoring.
