+++ tests/log/logtest.cpp
@@ -31,91 +31,7 @@
 // all calls to wxLogXXX() functions from this file will use this log component
 #define wxLOG_COMPONENT "test"
 
-// ----------------------------------------------------------------------------
-// test loggers
-// ----------------------------------------------------------------------------
-
-// base class for all test loggers which simply store all logged messages for
-// future examination in the test code
-class TestLogBase : public wxLog
-{
-public:
-    TestLogBase() { }
-
-    const wxString& GetLog(wxLogLevel level) const
-    {
-        return m_logs[level];
-    }
-
-    const wxLogRecordInfo& GetInfo(wxLogLevel level) const
-    {
-        return m_logsInfo[level];
-    }
-
-    void Clear()
-    {
-        for ( unsigned n = 0; n < WXSIZEOF(m_logs); n++ )
-        {
-            m_logs[n].clear();
-            m_logsInfo[n] = wxLogRecordInfo();
-        }
-    }
-
-protected:
-    wxString m_logs[wxLOG_Trace + 1];
-    wxLogRecordInfo m_logsInfo[wxLOG_Trace + 1];
-
-    wxDECLARE_NO_COPY_CLASS(TestLogBase);
-};
-
-// simple log sink which just stores the messages logged for each level
-class TestLog : public TestLogBase
-{
-public:
-    TestLog() { }
-
-protected:
-    virtual void DoLogRecord(wxLogLevel level,
-                             const wxString& msg,
-                             const wxLogRecordInfo& info) override
-    {
-        m_logs[level] = msg;
-        m_logsInfo[level] = info;
-    }
-
-private:
-    wxDECLARE_NO_COPY_CLASS(TestLog);
-};
-
-// ----------------------------------------------------------------------------
-// test class
-// ----------------------------------------------------------------------------
-
-class LogTestCase
-{
-public:
-    LogTestCase();
-    ~LogTestCase();
-
-protected:
-    TestLog *m_log;
-    wxLog *m_logOld;
-    bool m_logWasEnabled;
-
-    wxDECLARE_NO_COPY_CLASS(LogTestCase);
-};
-
-LogTestCase::LogTestCase()
-{
-    m_logOld = wxLog::SetActiveTarget(m_log = new TestLog);
-    m_logWasEnabled = wxLog::EnableLogging();
-}
-
-LogTestCase::~LogTestCase()
-{
-    delete wxLog::SetActiveTarget(m_logOld);
-    wxLog::EnableLogging(m_logWasEnabled);
-}
+#include "testlog.h"
 
 TEST_CASE_METHOD(LogTestCase, "wxLog::Functions", "[log]")
 {
